Relocate custom OAuth hooks
