--- apps/meteor/app/apple/client/index.ts
@@ -1,4 +1,0 @@
-import { CustomOAuth } from '../../../client/lib/customOAuth/CustomOAuth';
-import { config } from '../lib/config';
-
-CustomOAuth.configureOAuthService('apple', config);
