--- apps/meteor/app/nextcloud/client/index.ts
@@ -1,1 +1,0 @@
-import './useNextcloud';
