+++ apps/meteor/client/views/root/AppLayout.tsx
@@ -2,6 +2,15 @@ import { useEffect, Suspense, useSyncExternalStore } from 'react';
 
 import DocumentTitleWrapper from './DocumentTitleWrapper';
 import PageLoading from './PageLoading';
+import { useAppleOAuth } from './hooks/customOAuth/useAppleOAuth';
+import { useCustomOAuth } from './hooks/customOAuth/useCustomOAuth';
+import { useDolphinOAuth } from './hooks/customOAuth/useDolphinOAuth';
+import { useDrupalOAuth } from './hooks/customOAuth/useDrupalOAuth';
+import { useGitHubEnterpriseOAuth } from './hooks/customOAuth/useGitHubEnterpriseOAuth';
+import { useGitLabOAuth } from './hooks/customOAuth/useGitLabOAuth';
+import { useNextcloudOAuth } from './hooks/customOAuth/useNextcloudOAuth';
+import { useTokenpassOAuth } from './hooks/customOAuth/useTokenpassOAuth';
+import { useWordPressOAuth } from './hooks/customOAuth/useWordPressOAuth';
 import { useCodeHighlight } from './hooks/useCodeHighlight';
 import { useEscapeKeyStroke } from './hooks/useEscapeKeyStroke';
 import { useGoogleTagManager } from './hooks/useGoogleTagManager';
@@ -9,16 +18,9 @@ import { useLoadMissedMessages } from './hooks/useLoadMissedMessages';
 import { useLoginViaQuery } from './hooks/useLoginViaQuery';
 import { useMessageLinkClicks } from './hooks/useMessageLinkClicks';
 import { useSettingsOnLoadSiteUrl } from './hooks/useSettingsOnLoadSiteUrl';
-import { useWordPressOAuth } from './hooks/useWordPressOAuth';
 import { useCorsSSLConfig } from '../../../app/cors/client/useCorsSSLConfig';
-import { useDolphin } from '../../../app/dolphin/client/hooks/useDolphin';
-import { useDrupal } from '../../../app/drupal/client/hooks/useDrupal';
 import { useEmojiOne } from '../../../app/emoji-emojione/client/hooks/useEmojiOne';
-import { useGitHubEnterpriseAuth } from '../../../app/github-enterprise/client/hooks/useGitHubEnterpriseAuth';
-import { useGitLabAuth } from '../../../app/gitlab/client/hooks/useGitLabAuth';
 import { useLivechatEnterprise } from '../../../app/livechat-enterprise/hooks/useLivechatEnterprise';
-import { useNextcloud } from '../../../app/nextcloud/client/useNextcloud';
-import { useTokenPassAuth } from '../../../app/tokenpass/client/hooks/useTokenPassAuth';
 import { useIframeLoginListener } from '../../hooks/iframe/useIframeLoginListener';
 import { useNotificationPermission } from '../../hooks/notification/useNotificationPermission';
 import { useAnalytics } from '../../hooks/useAnalytics';
@@ -26,7 +28,6 @@ import { useAnalyticsEventTracking } from '../../hooks/useAnalyticsEventTracking
 import { useAutoupdate } from '../../hooks/useAutoupdate';
 import { useLoadRoomForAllowedAnonymousRead } from '../../hooks/useLoadRoomForAllowedAnonymousRead';
 import { appLayout } from '../../lib/appLayout';
-import { useCustomOAuth } from '../../sidebar/hooks/useCustomOAuth';
 import { useRedirectToSetupWizard } from '../../startup/useRedirectToSetupWizard';
 
 const AppLayout = () => {
@@ -50,12 +51,13 @@ const AppLayout = () => {
 	useRedirectToSetupWizard();
 	useSettingsOnLoadSiteUrl();
 	useLivechatEnterprise();
-	useNextcloud();
-	useGitLabAuth();
-	useGitHubEnterpriseAuth();
-	useDrupal();
-	useDolphin();
-	useTokenPassAuth();
+	useNextcloudOAuth();
+	useGitLabOAuth();
+	useGitHubEnterpriseOAuth();
+	useDrupalOAuth();
+	useDolphinOAuth();
+	useTokenpassOAuth();
+	useAppleOAuth();
 	useWordPressOAuth();
 	useCustomOAuth();
 	useCorsSSLConfig();
