+++ apps/meteor/client/views/root/hooks/customOAuth/useAppleOAuth.ts
@@ -1,0 +1,8 @@
+import { config } from '../../../../../app/apple/lib/config';
+import { CustomOAuth } from '../../../../lib/customOAuth/CustomOAuth';
+
+/* const Apple =*/ CustomOAuth.configureOAuthService('apple', config);
+
+export const useAppleOAuth = () => {
+	// do nothing
+};
