+++ apps/meteor/client/views/root/hooks/customOAuth/useDrupalOAuth.ts
@@ -2,7 +2,7 @@ import type { OauthConfig } from '@rocket.chat/core-typings';
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
-import { CustomOAuth } from '../../../../client/lib/customOAuth/CustomOAuth';
+import { CustomOAuth } from '../../../../lib/customOAuth/CustomOAuth';
 
 // Drupal Server CallBack URL needs to be http(s)://{rocketchat.server}[:port]/_oauth/drupal
 // In RocketChat -> Administration the URL needs to be http(s)://{drupal.server}/
@@ -25,7 +25,7 @@ const config = {
 
 const Drupal = CustomOAuth.configureOAuthService('drupal', config);
 
-export const useDrupal = () => {
+export const useDrupalOAuth = () => {
 	const drupalUrl = useSetting('API_Drupal_URL') as string;
 
 	useEffect(() => {
