+++ apps/meteor/client/views/root/hooks/customOAuth/useGitHubEnterpriseOAuth.ts
@@ -2,7 +2,7 @@ import type { OauthConfig } from '@rocket.chat/core-typings';
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
-import { CustomOAuth } from '../../../../client/lib/customOAuth/CustomOAuth';
+import { CustomOAuth } from '../../../../lib/customOAuth/CustomOAuth';
 
 // GitHub Enterprise Server CallBack URL needs to be http(s)://{rocketchat.server}[:port]/_oauth/github_enterprise
 // In RocketChat -> Administration the URL needs to be http(s)://{github.enterprise.server}/
@@ -20,7 +20,7 @@ const config = {
 
 const GitHubEnterprise = CustomOAuth.configureOAuthService('github_enterprise', config);
 
-export const useGitHubEnterpriseAuth = () => {
+export const useGitHubEnterpriseOAuth = () => {
 	const githubApiUrl = useSetting('API_GitHub_Enterprise_URL') as string;
 
 	useEffect(() => {
